Update exercise1 copy 2.c

+++ C/Embedded-Systems-Programming-TX00EX76-3001/Oma-tasks/Lab04/exercise1 copy 2.c
@@ -31,7 +31,7 @@
 typedef struct ledStatus
 {
     bool ledState[3];
-    uint8_t brightness;
+    uint16_t brightness;
 } ledStatus;
 
 typedef struct queueTrain
